refactor: move collapse_graphs to io._dot
