+++ src/expertsystem/reaction/__init__.py
@@ -235,120 +235,6 @@ class Result:
                         intermediate_states.add(particle)
         return intermediate_states
 
-    def get_particle_graphs(self) -> List[StateTransitionGraph[Particle]]:
-        """Strip `list` of `.StateTransitionGraph` s of the spin projections.
-
-        Extract a `list` of `.StateTransitionGraph` instances with only
-        particles on the edges.
-
-        .. seealso:: :doc:`/usage/visualize`
-        """
-        inventory: List[StateTransitionGraph[Particle]] = list()
-        for transition in self.transitions:
-            if any(
-                transition.compare(
-                    other, edge_comparator=lambda e1, e2: e1[0] == e2
-                )
-                for other in inventory
-            ):
-                continue
-            new_edge_props = dict()
-            for edge_id in transition.topology.edges:
-                edge_props = transition.get_edge_props(edge_id)
-                if edge_props:
-                    new_edge_props[edge_id] = edge_props[0]
-            inventory.append(
-                StateTransitionGraph[Particle](
-                    topology=transition.topology,
-                    node_props={
-                        i: node_props
-                        for i, node_props in zip(
-                            transition.topology.nodes,
-                            map(
-                                transition.get_node_props,
-                                transition.topology.nodes,
-                            ),
-                        )
-                        if node_props
-                    },
-                    edge_props=new_edge_props,
-                )
-            )
-        inventory = sorted(
-            inventory,
-            key=lambda g: [
-                g.get_edge_props(i).mass
-                for i in g.topology.intermediate_edge_ids
-            ],
-        )
-        return inventory
-
-    def collapse_graphs(
-        self,
-    ) -> List[StateTransitionGraph[ParticleCollection]]:
-        def merge_into(
-            graph: StateTransitionGraph[Particle],
-            merged_graph: StateTransitionGraph[ParticleCollection],
-        ) -> None:
-            if (
-                graph.topology.intermediate_edge_ids
-                != merged_graph.topology.intermediate_edge_ids
-            ):
-                raise ValueError(
-                    "Cannot merge graphs that don't have the same edge IDs"
-                )
-            for i in graph.topology.edges:
-                particle = graph.get_edge_props(i)
-                other_particles = merged_graph.get_edge_props(i)
-                if particle not in other_particles:
-                    other_particles += particle
-
-        def is_same_shape(
-            graph: StateTransitionGraph[Particle],
-            merged_graph: StateTransitionGraph[ParticleCollection],
-        ) -> bool:
-            if graph.topology.edges != merged_graph.topology.edges:
-                return False
-            for edge_id in (
-                graph.topology.incoming_edge_ids
-                | graph.topology.outgoing_edge_ids
-            ):
-                edge_prop = merged_graph.get_edge_props(edge_id)
-                if len(edge_prop) != 1:
-                    return False
-                other_particle = next(iter(edge_prop))
-                if other_particle != graph.get_edge_props(edge_id):
-                    return False
-            return True
-
-        particle_graphs = self.get_particle_graphs()
-        inventory: List[StateTransitionGraph[ParticleCollection]] = list()
-        for graph in particle_graphs:
-            append_to_inventory = True
-            for merged_graph in inventory:
-                if is_same_shape(graph, merged_graph):
-                    merge_into(graph, merged_graph)
-                    append_to_inventory = False
-                    break
-            if append_to_inventory:
-                new_edge_props = {
-                    edge_id: ParticleCollection(
-                        {graph.get_edge_props(edge_id)}
-                    )
-                    for edge_id in graph.topology.edges
-                }
-                inventory.append(
-                    StateTransitionGraph[ParticleCollection](
-                        topology=graph.topology,
-                        node_props={
-                            i: graph.get_node_props(i)
-                            for i in graph.topology.nodes
-                        },
-                        edge_props=new_edge_props,
-                    )
-                )
-        return inventory
-
 
 @attr.s
 class ProblemSet:
