+++ tests/unit/reaction/test_solving.py
@@ -1,6 +1,5 @@
 # pylint: disable=no-self-use
 
-from expertsystem.particle import ParticleCollection
 from expertsystem.reaction import Result
 
 
@@ -11,40 +10,3 @@ class TestResult:
         result = jpsi_to_gamma_pi_pi_helicity_solutions
         intermediate_particles = result.get_intermediate_particles()
         assert intermediate_particles.names == {"f(0)(1500)", "f(0)(980)"}
-
-    def test_get_particle_graphs(
-        self,
-        particle_database: ParticleCollection,
-        jpsi_to_gamma_pi_pi_helicity_solutions: Result,
-    ):
-        pdg = particle_database
-        result = jpsi_to_gamma_pi_pi_helicity_solutions
-        particle_graphs = result.get_particle_graphs()
-        assert len(particle_graphs) == 2
-        assert particle_graphs[0].get_edge_props(3) == pdg["f(0)(980)"]
-        assert particle_graphs[1].get_edge_props(3) == pdg["f(0)(1500)"]
-        assert len(particle_graphs[0].topology.edges) == 5
-        for edge_id in range(-1, 3):
-            assert particle_graphs[0].get_edge_props(
-                edge_id
-            ) is particle_graphs[1].get_edge_props(edge_id)
-
-    def test_collapse_graphs(
-        self,
-        jpsi_to_gamma_pi_pi_helicity_solutions: Result,
-        particle_database: ParticleCollection,
-    ):
-        pdg = particle_database
-        result = jpsi_to_gamma_pi_pi_helicity_solutions
-        particle_graphs = result.get_particle_graphs()
-        assert len(particle_graphs) == 2
-        collapsed_graphs = result.collapse_graphs()
-        assert len(collapsed_graphs) == 1
-        graph = next(iter(collapsed_graphs))
-        edge_id = next(iter(graph.topology.intermediate_edge_ids))
-        f_resonances = pdg.filter(
-            lambda p: p.name in ["f(0)(980)", "f(0)(1500)"]
-        )
-        intermediate_states = graph.get_edge_props(edge_id)
-        assert isinstance(intermediate_states, ParticleCollection)
-        assert intermediate_states == f_resonances
